health_tracker/requirements.txt
Flask==2.0.1
Flask-MySQLdb==0.2.0
Werkzeug==2.0.1
mysqlclient==2.1.0